Refactor: auth コンテキストを導入
auth 関連ファイルを contexts/auth/application/ に移動

## 変更内容
- admin.ts, client.ts, login.ts, roles.ts を移動
- import パスを更新

## 備考
- auth は DDD らしい構造ではないが、機能として独立しているため application 層のみで構成
- prisma-user.repository は auth 外からも参照されているため shared に移動予定（Phase 3 残作業）
