+++ admin/src/app/api/auth/setup-password/route.ts
@@ -1,6 +1,6 @@
 import "server-only";
 import { NextResponse } from "next/server";
-import { createClient } from "@/server/auth/client";
+import { createClient } from "@/server/contexts/auth/application/client";
 
 export async function POST(request: Request) {
   try {
