+++ admin/src/app/api/logout/route.ts
@@ -1,4 +1,4 @@
-import { createClient } from "@/server/auth/client";
+import { createClient } from "@/server/contexts/auth/application/client";
 import { NextResponse } from "next/server";
 
 export const runtime = "nodejs";
