+++ admin/src/app/api/users/invite/route.ts
@@ -1,7 +1,7 @@
 import "server-only";
 import { NextResponse } from "next/server";
-import { requireRole } from "@/server/auth/roles";
-import { createAdminClient } from "@/server/auth/admin";
+import { requireRole } from "@/server/contexts/auth/application/roles";
+import { createAdminClient } from "@/server/contexts/auth/application/admin";
 
 export async function POST(request: Request) {
   try {
