+++ admin/src/server/contexts/auth/application/roles.ts
@@ -1,4 +1,4 @@
-import { createClient } from "@/server/auth/client";
+import { createClient } from "@/server/contexts/auth/application/client";
 import type { UserRole } from "@prisma/client";
 import { prisma } from "@/server/lib/prisma";
 import { PrismaUserRepository } from "@/server/repositories/prisma-user.repository";
